Changed variable type names. Added SPI skeleton.

+++ Hall.c
@@ -5,6 +5,8 @@
 */
 
 #include <LPC214x.H>
+#include "Type.h"
+
 #include "Hall.h"
 
 /*
@@ -14,8 +16,8 @@
 */
 #define RES 1
 
-unsigned int last_capture1 = 0, signal_freq1 = 0;
-unsigned int last_capture2 = 0, signal_freq2 = 0;
+uint32 last_capture1 = 0, signal_freq1 = 0;
+uint32 last_capture2 = 0, signal_freq2 = 0;
 
 void T1isr(void)    __irq;
 
@@ -29,12 +31,12 @@ void capture_init(void)
     T1CCR = 0x00000005 | 0x00000028;    //Capture on rising edge of channel [0 + 1] + interrupt
     T1TCR = 0x00000001;                 //Enable timer
 
-    VICVectAddr4 = (unsigned) T1isr;    //Set the timer ISR vector address
+    VICVectAddr4 = (uint32) T1isr;      //Set the timer ISR vector address
     VICVectCntl4 = 0x00000020 | 5;      //Set channel for TIMER1
     VICIntEnable |= (1 << 5);           //Enable the interrupt TIMER1
 }
 
-unsigned int hall_get(int nr)
+uint32 hall_get(uint8 nr)
 {
     if (nr == 1) return signal_freq1;   //Rotations per second for left motor
     if (nr == 2) return signal_freq2;   //Rotations per second for right motor
@@ -42,7 +44,7 @@ unsigned int hall_get(int nr)
     return 0;
 }
 
-void hall_update(int nr, unsigned int new_value)
+void hall_update(uint8 nr, uint32 new_value)
 {
     if (nr == 1) signal_freq1 = new_value;
     if (nr == 2) signal_freq2 = new_value;
@@ -50,7 +52,7 @@ void hall_update(int nr, unsigned int new_value)
     return;
 }
 
-void hall_reset(int nr)
+void hall_reset(uint8 nr)
 {
     switch (nr)
     {
@@ -66,15 +68,15 @@ void hall_reset(int nr)
     }
 }
 
-unsigned int compute_freq(unsigned int current, unsigned int last)
+uint32 compute_freq(uint32 current, uint32 last)
 {
-    unsigned int den;
+    uint32 den;
     den = current - last;
 
     return (1000000 * RES / den);
 }
 
-unsigned int hall_now(int nr)
+uint32 hall_now(uint8 nr)
 {
     if (nr == 1)
         return compute_freq(T1TC, last_capture1);
@@ -87,7 +89,7 @@ unsigned int hall_now(int nr)
 
 void T1isr(void)	__irq
 {
-    unsigned int value;
+    uint32 value;
 
     if (T1IR & (1 << 4))                        //Capture channel 0
     {
