+++ Hall.h
@@ -16,25 +16,25 @@ void capture_init(void);
     Read the motor speed (rotations / sec)
     for each motor (nr - 1, 2)
 */
-unsigned int hall_get(int nr);
+uint32 hall_get(uint8 nr);
 
 /*
     Change speed manually, from outside the capture ISR
     This function is used in a task for Hall sensor timeout
 */
-void hall_update(int nr, unsigned int new_value);
+void hall_update(uint8 nr, uint32 new_value);
 
 /*
     Reset the measurements to 0 for
     1 - left motor, 2 - right motor, 3 - both motors
 */
-void hall_reset(int nr);
+void hall_reset(uint8 nr);
 
 /*
     Computes a virtual speed using the difference between the
     current Timer1 counter (T1TC) and the last captured value
     nr - 1 or 2 [motor 1 or 2]
 */
-unsigned int hall_now(int nr);
+uint32 hall_now(uint8 nr);
 
 #endif  // __HALL_H
